# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sanktionssök API",
    "description": "REST API för att söka mot konsoliderade sanktionslistor (EU, FN, US OFAC) samt PEP-listor.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokal utvecklingsserver"
    }
  ],
  "paths": {
    "/api/search": {
      "get": {
        "summary": "Sök i sanktionslistor",
        "description": "Sök efter personer eller organisationer baserat på namn, alias eller ID i EU:s, FN:s, USA:s och PEP-listor. Kräver antingen en inloggad session eller en API-token med read-scope (issue #36).",
        "security": [{ "cookieAuth": [] }, { "apiTokenAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Sökord (namn, alias, passnummer)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filtrera på källa (t.ex. EU, UN, US, PEP). Separera med kommatecken.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filtrera på typ",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["individual", "entity", "vessel", "aircraft"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max antal träffar (standard 20, min 0, max 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "includeDelisted",
            "in": "query",
            "description": "Om 'true', inkludera avlistade (soft-deleted) poster i träfflistan. Standard är att endast visa aktiva poster.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "description": "Lägsta matchningspoäng (0-100) för att en träff ska tas med. Standard 65.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100,
              "default": 65
            }
          },
          {
            "name": "dob",
            "in": "query",
            "description": "Födelsedatum (helt eller delvis, t.ex. ett årtal) som boostar — men inte krävs av — poster vars birthDates innehåller värdet. Inte ett hårt filter: en post utan matchande datum exkluderas aldrig enbart av detta.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "1952-10-07"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sökresultat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Felaktig sökförfrågan (t.ex. saknat sökord)"
          },
          "401": {
            "description": "Ingen giltig session eller API-token"
          },
          "403": {
            "description": "API-token saknar read-scope"
          }
        }
      }
    },
    "/api/sanctions/{id}": {
      "get": {
        "summary": "Hämta en specifik sanktionspost",
        "description": "Hämta detaljerad information om en sanktionerad person eller organisation baserat på dess unika ID. Kräver antingen en inloggad session eller en API-token med read-scope (issue #36).",
        "security": [{ "cookieAuth": [] }, { "apiTokenAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unikt ID för sanktionsposten (t.ex. EU-1234, US-SDN-999)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hittad post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanctionRecord"
                }
              }
            }
          },
          "401": {
            "description": "Ingen giltig session eller API-token"
          },
          "403": {
            "description": "API-token saknar read-scope"
          },
          "404": {
            "description": "Hittade inte posten"
          }
        }
      }
    },
    "/api/sanctions/{id}/versions": {
      "get": {
        "summary": "Hämta versionshistorik för en sanktionspost",
        "description": "Listar hela ändringshistoriken (skapad/uppdaterad/avlistad/relistad) för en post, nyaste först. Kräver antingen en inloggad session eller en API-token med read-scope (issue #36).",
        "security": [{ "cookieAuth": [] }, { "apiTokenAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unikt ID för sanktionsposten",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Versionshistorik, nyaste först",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/RecordVersion" }
                }
              }
            }
          },
          "404": { "description": "Hittade inte posten" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Session-cookie satt av POST /api/auth/verify-otp efter lyckad e-post-OTP-inloggning. Krävs av alla /api/*-rutter förutom auth-rutterna själva och /api/admin/tokens/* (endast admin-session, se requireAdmin)."
      },
      "apiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sanc_<random>",
        "description": "API-token skapad via POST /api/admin/tokens, skickad som 'Authorization: Bearer sanc_...'. Stödjer granulära scopes (sanctions:read, custom:read/write, overrides:read/write, decisions:read/write, imports:read/write, system:read) samt bakåtkompatibla 'read'/'write'-paket (issue #36, #221)."
      }
    },
    "schemas": {
...
      "SearchResponse": {
        "type": "object",
        "description": "Faktiskt svar från GET /api/search (src/api/index.ts) — inte en bar array, till skillnad från hur det tidigare stod i denna spec.",
        "required": ["results", "totalMatches", "truncated"],
        "properties": {
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ScoredResult" }
          },
          "totalMatches": {
            "type": "integer",
            "description": "Antal träffar innan 'limit' tillämpades."
          },
          "truncated": {
            "type": "boolean",
            "description": "true om totalMatches är fler än vad 'results' innehåller, dvs fler träffar finns än vad som returnerades."
          }
        }
      },
      "ScoredResult": {
        "description": "En sanktionspost så som den returneras av GET /api/search, med matchningsinformation utöver den lagrade posten.",
        "allOf": [
          { "$ref": "#/components/schemas/SanctionRecord" },
          {
            "type": "object",
            "required": ["score", "matchedAlias", "overriddenFields"],
            "properties": {
              "score": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100,
                "description": "Matchningspoäng för denna sökning (fuzzy namnmatchning, eller 100 vid exakt pass-/ID-träff)."
              },
              "matchedAlias": {
                "type": "string",
                "description": "Vilket namn i names[] (primärt eller alias) som gav den högsta poängen."
              },
              "overriddenFields": {
                "type": "array",
                "items": { "type": "string" },
                "description": "Fältnamn som lokalt korrigerats via /api/overrides och sammanslagits in i denna post."
              }
            }
          }
        ]
      },
...
      "SanctionRecord": {
        "type": "object",
        "required": ["id", "source", "type", "names", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "example": "EU-1234"
          },
          "source": {
            "type": "string",
            "enum": ["EU", "UN", "US", "UK", "PEP", "CUSTOM"],
            "example": "EU"
          },
          "type": {
            "type": "string",
            "enum": ["individual", "entity", "vessel", "aircraft"],
            "example": "individual"
          },
          "names": {
            "type": "array",
            "description": "Every name/alias for this entity, primary name first (issue #46 — the sole source of truth; no separate flat primaryName/aliases fields).",
            "items": {
              "type": "object",
              "required": ["wholeName", "strong"],
              "properties": {
                "wholeName": { "type": "string", "example": "Vladimir Vladimirovich Putin" },
                "firstName": { "type": "string" },
                "middleName": { "type": "string" },
                "lastName": { "type": "string" },
                "strong": { "type": "boolean", "description": "Authoritative name vs. a weaker variant, per the source." },
                "language": { "type": "string" },
                "title": { "type": "string" },
                "function": { "type": "string" }
              }
            },
            "example": [
              { "wholeName": "Vladimir Vladimirovich Putin", "strong": true },
              { "wholeName": "Valodia Putin", "strong": false }
            ]
          },
          "birthDates": {
            "type": "array",
            "description": "Structured, precision-aware birth dates (exact date, year-only, year range, or circa/approximate).",
            "items": {
              "type": "object",
              "properties": {
                "raw": { "type": "string", "example": "1952-10-07" },
                "year": { "type": "integer" },
                "month": { "type": "integer" },
                "day": { "type": "integer" },
                "yearRangeFrom": { "type": "integer" },
                "yearRangeTo": { "type": "integer" },
                "circa": { "type": "boolean" },
                "place": { "type": "string" },
                "city": { "type": "string" },
                "countryIso2": { "type": "string" }
              }
            }
          },
          "placesOfBirth": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Leningrad, USSR"]
          },
          "citizenships": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Russia"]
          },
          "identifications": {
            "type": "array",
            "description": "Passports, national IDs, and similar documents — structured, including the source's own reliability flags (CLAUDE.md §6: never present a known-false/expired/revoked document as good data).",
            "items": {
              "type": "object",
              "required": ["number"],
              "properties": {
                "number": { "type": "string", "example": "123456" },
                "typeCode": { "type": "string" },
                "typeDescription": { "type": "string", "example": "Diplomatic Passport" },
                "countryIso2": { "type": "string" },
                "issuedBy": { "type": "string" },
                "knownFalse": { "type": "boolean" },
                "knownExpired": { "type": "boolean" },
                "reportedLost": { "type": "boolean" },
                "revokedByIssuer": { "type": "boolean" },
                "diplomatic": { "type": "boolean" }
              }
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "street": { "type": "string" },
                "city": { "type": "string" },
                "country": { "type": "string" },
                "fullAddress": { "type": "string" }
              }
            }
          },
          "sanctionReason": {
            "type": "string",
            "example": "Involvement in destabilizing Ukraine"
          },
          "legalBasis": {
            "type": "string",
            "example": "Regulation (EU) No 269/2014"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["active", "delisted"],
            "description": "Soft-delete status. A missing value should be treated as 'active'.",
            "example": "active"
          },
          "listedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record first appeared in an import."
          },
          "delistedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record most recently went missing from an import."
          },
          "overriddenFields": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Field names currently overridden by a local correction (issue #35). Not part of the persisted document — computed on every read by merging overrides/{id} onto the stored record, so the caller can tell official source data apart from a local correction. Empty when no override exists for this record.",
            "example": ["sanctionReason"]
          }
        }
      },
...
      "RecordVersion": {
        "type": "object",
        "required": ["importId", "changedAt", "changeType", "record"],
        "properties": {
          "importId": { "type": "string" },
          "changedAt": { "type": "string", "format": "date-time" },
          "changeType": {
            "type": "string",
            "enum": ["created", "updated", "delisted", "relisted"]
          },
          "record": { "$ref": "#/components/schemas/SanctionRecord" }
        }
      },
...
    }
  }
}